Correctly erase userdata partition
We have two different cases for the erase of our userdata partition. For the first case
the userdata partition is left alone without any image we need to loop-mount later on and
we can do a complete format to get it into a sane state. For the other case we can just
erase all files but have to keep those images we have to loop-mount. Previously this erase
didn't respected hidden files, directories or special files. This is now working as
expected.

Fixes LP #1419928

Change-Id: Iaf697f59a7e8cfdf225fd4d4ce3f34065a2eb9c9
Signed-off-by: Simon Busch <simon.busch@canonical.com>