Don't erase certain files on userdata reset
There are some files we want to keep on userdata reset like when we have last updated the
device with an OTA or relevant machine IDs. Those files are kept in a whitelist which are
backuped before doing the erase and restored afterwards. The whitelist is kept in a
separate file within the recovery for easier maintenance.

Currently the whitelist includes files to indicate when the last OTA update was
installed.

This also fixes the system-image-upgrader to not touch /data/.last_update when no update
is installed.

Change-Id: Idcd5590f97e76593d5a58bf5cad71cc39caa24f5
Signed-off-by: Simon Busch <simon.busch@canonical.com>