Redesign the recovery visuals.
Now the aubergine 'paper' background is used, with a centered
white text Ubuntu logo and an indeterminate progress bar below
that.

Since the background is no longer a solid color, it can't just
be tiled like the old one.  Which means it needs to be properly
cropped / resized for each device.  And the logo needs to be
likewise remade when that happens (because devices can't handle
transparency, we have to fake it by creating a logo with the
centered part of the background built in that we just blit on
top of the background so that it's seamless).  Toward that end,
I've added the source images to recreate the images for each
device.  Specifically, full-background.jpg is the full-size
default Ubuntu background and logo-overlay.png is the
transparent logo that just needs to be centered on the
background.

Since the background and logo need to be device-specific now
(or at least, screen-resolution-specific), I've picked mako as
the reference device.  Other devices can use res/image overrides
in their device folders.

Change-Id: Idf4a6395d46319c78f54f2f17b8b17e31908da9c
Signed-off-by: Michael Terry <michael.terry@canonical.com>